Refactored peers api and folder structure.
* Removed peer related endpoints from the lightning package and moved it to it's own package.
* Removed the content of the old peer package. The old content was only used to assist with channel open.
* API now sending string values of connected status and reconnection setting.
