+++ internal/nodes/routes.go
@@ -2,11 +2,10 @@ package nodes
 
 import (
 	"fmt"
-	"net/http"
-	"strconv"
-
 	"github.com/gin-gonic/gin"
 	"github.com/jmoiron/sqlx"
+	"net/http"
+	"strconv"
 
 	"github.com/lncapital/torq/internal/core"
 	"github.com/lncapital/torq/pkg/server_errors"
@@ -43,19 +42,7 @@ type NodeWalletBalance struct {
 	ReservedBalanceAnchorChan int64 `json:"reservedBalanceAnchorChan"`
 }
 
-type PeerNode struct {
-	NodeId           int                         `json:"nodeId" db:"node_id"`
-	Alias            string                      `json:"peerAlias" db:"alias"`
-	PublicKey        string                      `json:"pubKey" db:"public_key"`
-	TorqNodeId       *int                        `json:"torqNodeId" db:"torq_node_id"`
-	TorqNodeAlias    *string                     `json:"torqNodeAlias" db:"torq_node_alias"`
-	Setting          *core.NodeConnectionSetting `json:"setting" db:"setting"`
-	ConnectionStatus *core.Status                `json:"connectionStatus" db:"connection_status"`
-	Address          *string                     `json:"address" db:"address"`
-}
-
 func RegisterNodeRoutes(r *gin.RouterGroup, db *sqlx.DB) {
-	r.GET("/:network/peers", func(c *gin.Context) { getAllPeersHandler(c, db) })
 	r.GET("/:network/nodes", func(c *gin.Context) { getNodesByNetworkHandler(c, db) })
 	r.DELETE(":nodeId", func(c *gin.Context) { removeNodeHandler(c, db) })
 }
@@ -85,19 +72,4 @@ func removeNodeHandler(c *gin.Context, db *sqlx.DB) {
 		return
 	}
 	c.JSON(http.StatusOK, map[string]interface{}{"message": fmt.Sprintf("Successfully deleted %v node(s).", count)})
-}
-
-func getAllPeersHandler(c *gin.Context, db *sqlx.DB) {
-	network, err := strconv.Atoi(c.Param("network"))
-	if err != nil {
-		server_errors.SendBadRequest(c, "Can't process network")
-		return
-	}
-	peerNodes, err := GetPeerNodes(db, core.Network(network))
-	if err != nil {
-		server_errors.WrapLogAndSendServerError(c, err, "Getting all Peer nodes.")
-		return
-	}
-
-	c.JSON(http.StatusOK, peerNodes)
 }
