+++ web/src/features/peers/peersCellRenderer.tsx
@@ -1,5 +1,5 @@
 import { ColumnMetaData } from "features/table/types";
-import { ConnectionStatus, NodeConnectionSetting, Peer } from "features/peers/peersTypes";
+import { Peer } from "features/peers/peersTypes";
 import DefaultCellRenderer from "features/table/DefaultCellRenderer";
 import useTranslations from "services/i18n/useTranslations";
 import TextCell from "components/table/cells/text/TextCell";
@@ -15,34 +15,30 @@ export default function peerCellRenderer(
 ): JSX.Element {
   const { t } = useTranslations();
 
-  if (column.key === "connectionStatus") {
-    return (
-      <TextCell
-        current={t.peersPage[ConnectionStatus[row.connectionStatus]]}
-        key={column.key.toString() + rowIndex}
-        totalCell={isTotalsRow}
-      />
-    );
-  }
-
-  if (column.key === "setting") {
-    return (
-      <TextCell
-        current={t.peersPage[NodeConnectionSetting[row.setting]]}
-        key={column.key.toString() + rowIndex}
-        totalCell={isTotalsRow}
-      />
-    );
-  }
-
   if (column.key === "peerAlias") {
     return (
       <PeersAliasCell
+        key={column.key.toString() + rowIndex}
         alias={row.peerAlias}
         peerNodeId={row.nodeId}
         torqNodeId={row.torqNodeId}
         connectionStatus={row.connectionStatus}
       />
+    );
+  }
+
+  if (column.key === "connectionStatus") {
+    return (
+      <TextCell
+        current={t.peersPage[row.connectionStatus]}
+        key={column.key.toString() + rowIndex}
+        totalCell={isTotalsRow}
+      />
+    );
+  }
+  if (column.key === "setting") {
+    return (
+      <TextCell current={t.peersPage[row.setting]} key={column.key.toString() + rowIndex} totalCell={isTotalsRow} />
     );
   }
